Use pseudo-build process
I'm not particularly proud of this solution as it requires manual
intervention to run tests.
Still searching for a way to very simply bundle the entrypoint functions
such that they are the only ones detected by GAS. The bundlers I've
tried thus far (esbuild, rollup) add too much boilerplate and expose
additional functions that GAS shouldn't care about.
This works for now, but I'll continue looking for a tool suited for the
above (perhaps creating my own if necessary).